Reconfigured project after finishing it
I was able to get the page looking like the example image using
margin and padding parameters, but figured later that having a
set width for each section would work better with resizing. I
decided to start the project from scratch, using some of the
css I had done before, but this time used a percentage of the vw
for my section widths and wrapped each one around a div to be able
to easily set the background color across the whole page. This
solution is the closest I was able to get to the example image.
My only issue is with the way info cards resize when making the
page smaller (text overlaps eventually, probably has to do with
the flex attribute).

I also replaced the information with images and text from my band.
Did it more as a joke than anything else, but helped me have an
actual reference for the content in the page.